scripts/requirements.txt
openai>=1.30.0
python-docx>=1.1.0
nbformat>=5.9.0
